fix: restructure suspense
